requirements.txt
# Core Dependencies
# REQUIRES PYTHON 3.10+ (Recommended: Python 3.12)
# Compatible with Gradio 5.x and 6.x
gradio>=6.0.0
anthropic>=0.34.0
mcp>=1.21.2  # FastMCP API required
httpx>=0.25.0

# Web Scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Database Access (for AACT clinical trials database)
psycopg2-binary>=2.9.0
asyncpg>=0.29.0  # For async PostgreSQL with connection pooling

# Configuration
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# RAG and Research Memory (LlamaIndex)
llama-index-core>=0.11.0
llama-index-vector-stores-chroma>=0.2.0
llama-index-embeddings-huggingface>=0.3.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
transformers>=4.30.0

# Development
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
